feat: Controller Advice을 이용한 예외 처리, CustomerErrorController를 통한 404 에러 등 다양한 에러 처리

+++ src/main/java/com/programmers/springweekly/controller/CustomerViewController.java
@@ -5,6 +5,7 @@ import com.programmers.springweekly.dto.customer.response.CustomerListResponse;
 import com.programmers.springweekly.dto.customer.response.CustomerResponse;
 import com.programmers.springweekly.service.CustomerService;
 import com.programmers.springweekly.util.validator.CustomerValidator;
+import java.util.NoSuchElementException;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -64,6 +65,12 @@ public class CustomerViewController {
 
     @GetMapping("/delete/{id}")
     public String deleteById(@PathVariable("id") UUID customerId) {
+        boolean isExistCustomerId = customerService.existById(customerId);
+
+        if (!isExistCustomerId) {
+            throw new NoSuchElementException("사용자가 삭제하려는 아이디 " + customerId + "는 없는 ID입니다.");
+        }
+
         customerService.deleteById(customerId);
 
         return "redirect:/view/customer/find";
