+++ src/main/java/com/programmers/springweekly/controller/VoucherViewController.java
@@ -5,6 +5,7 @@ import com.programmers.springweekly.dto.voucher.response.VoucherListResponse;
 import com.programmers.springweekly.dto.voucher.response.VoucherResponse;
 import com.programmers.springweekly.service.VoucherService;
 import com.programmers.springweekly.util.validator.VoucherValidator;
+import java.util.NoSuchElementException;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -61,6 +62,12 @@ public class VoucherViewController {
 
     @GetMapping("/delete/{id}")
     public String deleteById(@PathVariable("id") UUID voucherId) {
+        boolean isExistVoucherId = voucherService.existById(voucherId);
+
+        if (!isExistVoucherId) {
+            throw new NoSuchElementException("사용자가 삭제하려는 바우처 " + voucherId + "는 없는 ID입니다.");
+        }
+
         voucherService.deleteById(voucherId);
 
         return "redirect:/view/voucher/find";
